Use microbe avatars in poker UI when petri dish mode is active
The poker game always rendered players as fish, even when the petri
world was selected. The active world type is now threaded from
TankView through TankPokerTab/PokerGame to the poker avatar
components, which render the player's genome with the same canvas
microbe renderer (drawMicrobe) used in the dish, so poker avatars
match the organisms on screen.

Additional UI fixes along the way:
- Deduplicate buildFishParams/renderPattern (PokerPlayer had a stale
  copy missing chevron and scales patterns, so those fish showed no
  pattern in player cards) into shared fishAvatarParts.tsx
- Replace blocking alert() calls in TankPokerTab with the existing
  non-blocking poker error notification
- World-aware labels: 'Microbe #N' avatar titles, leaderboard column
  header, and poker loading copy in petri mode

Verified in-browser: petri mode shows microbe cells (nucleus/cilia/
flagella) in poker player cards, tank mode still shows SVG fish.
npm run lint, build, and vitest all pass.
